:horse: created separate interface for ponto user access
